Simplify xtask
lib/bin/test separation isn't really needed.
