+++ xtask/src/pre_cache.rs
@@ -6,12 +6,12 @@ use std::{
 use anyhow::Result;
 use xshell::rm_rf;
 
-pub struct PreCacheCmd;
+pub(crate) struct PreCacheCmd;
 
 impl PreCacheCmd {
     /// Cleans the `./target` dir after the build such that only
     /// dependencies are cached on CI.
-    pub fn run(self) -> Result<()> {
+    pub(crate) fn run(self) -> Result<()> {
         let slow_tests_cookie = Path::new("./target/.slow_tests_cookie");
         if !slow_tests_cookie.exists() {
             panic!("slow tests were skipped on CI!")
